Quick fixes 2018 (#270)
* made a hot fix for shift pagination to fit the users view

* Links Electrical Queue in sidebar just like Structural Queue.

* Add placeholder text to reason field for org timeline entries.

* Hide status and downtime from org pages for non-building orgs.

Also show a dash for the status of non-building orgs on the org index, when
shown.

* List each shift's type on shifts index when a type is not selected for filtering.

* Show a message when there are no results for a search query.

Also format existing code with consistent indentation.

* Simplify code for formatting time durations.

This fixes incorrect formatting, e.g., '0:7' on the structural and electrical
queue pages.

* Adds remaining downtime amount to org downtime detail page.

Also fixes responsive table columns.

* Add @merichar's Andrew ID to the FAQ so people know how to look her up.

* Madelyn Miller retired as director of EH&S, effective 2017-12-31.

Melanie Lucht is succeeding her on an interim basis.
https://www.cmu.edu/piper/news/archives/2017/may/may-4/personal-mention.html#miller

* Add placeholder message for receiving participant in the charge form.

Also fixes error in the form's pre-population of receiving participants on
existing charges.

* Format dollar values as currency.

* Link to each organization's charges from overall charges list.

Instead of linking directly to the organization page.

* Split shift lists into current/upcoming and past.

This makes finding the next few or last few shifts straightforward, rather than
having to page through to find the current time.  Highlight current shifts
according to checked-in status.

Since there are now two lists on the page, reduce the pagination from 20 to 15.