Remove Pagy dependency in participant and tools controllers
Pagy was not respecting pagination_size parameters when tabulator tables were getting data, causing each batch to only contain 20 entries. For the participant and tools controllers, this leads to 20-30s loading times that render the tables difficult to use for that time.

Pagy has been replaced by manually checking for "page" and "size" params, and returning the correct entries accordingly.